+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: earnaud <earnaud@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/31 02:24:50 by earnaud           #+#    #+#             */
-/*   Updated: 2021/06/17 11:46:28 by earnaud          ###   ########.fr       */
+/*   Updated: 2021/06/22 15:06:28 by earnaud          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ t_stacks	*convert_stack(int argc, char **argv);
 int			check_error(int argc, char **argv);
 
 int			stack_size(long *stack);
-void		find_min_max(long *stack, int *min_max);
+void		find_min_max(long *stack, long *min_max);
 long		find_min(long *stack);
 long		where_in(long *stack, long nbr);
 long		find_next(long *stack, long current);
